feat: support cleaner install-addons map
The original format was ensuring that pattern is supported, but as i've
seen more 'libraries', and since i'm going to be consuming my own, i'd
rather lean harder into the convention, which seems to be pretty
consistent.

Mostly i'm referring to where people put the `addons` dir in their
`addon` lib, and whether the names match the repo name, etc.

The assumption here is that the name used in the :keyword matches the
addons/<dir-name> that needs to be symlinked.

repo-id is a concept that means <user-or-org-name>/<repo-name>, but as a
string. tho it could also be a namespace keyword....hmmmmm.