feat: working navi pause menu
Uses a popup panel and ensured by Navi on autoload. Working great from
dino's main menu.

Testing and what goes where is a wip. Definitely want to write a few
tests for this, as it was kind of a pain.

The Navi node creates a canvasLayer node to attach the in-lined
PausePopup, so that it takes over the screen/camera properly (otherwise
it was landing at (0, 0)).

The thinking is that we want to support several versions of this being
consumed:
- the game jam case where we just don't do anything and the pause menu
is consumed
- the mid-range game that can re-use the scene type OR the script to get
a decently branded pause-panel with minimal effort
- the end-game that can get away with using Navi yet impling it's own
pause menu

Perhaps we should write tests for each? But then again, creating our own
test scene and testing that might provide coverage for all three.