reptile: basic 2x2 and 3x3 autotile asset, tileset, tilemap
Reptile now provides a starter tilemap asset, which can be used to
quickly doodle a tilemap with auto-tiling. For more serious usage, the
tileset resources can be 'made-unique' in godot, and the source png can
be updated via 'edit-dependencies' (once you've created a new image to
point to). Note that editing deps on a resource typically requires a
reload-current-project.