feat: refactor arrow collision to area2d
not sure why the rigidBody body_entered never fires here... seems odd,
but maybe that's not the intended usage. We continue using the rigid
body, but disable the collision layers - it's really just for the
apply_impulse so we don't have to move the arrow ourselves. A new area2d
added to the arrow is used to detect a collision, apply 'hit' to any
bodies that support it, and kill the arrow.