feat: LeafDrops room, improved room offset gluing
The Area2d must still be at the room origin, but the collisionShape's
offset is now taken into account.

Nice to remove this bit of logic from the runner, but it feels like we
might prefer an entrance-position/exit-position api instead - then some
logic for getting those to align. That'd be a nice way to break us away
from the origin a bit.