feat: Add custom dir-prefix to install system
The Addons system may be in the process of being removed but as long as the
commands exist here, other users may want the repos cloned in a different
location than directly under their home directory. This change makes the
addons-* and install-script-templates commands take variadic additional args
from the command line, each word of which will be interpreted as a part of the
path under the user's $HOME directory in which they want the repos to be cloned,
or in the case of the script_templates, where the respective repo can be found
on the local machine.