feat: Make the git status output for addons better
Rather than the full message for the git status output, which gets a bit lost
when being emitted for each addon repo, use git status -sb to show the short
format with the branch and tracking information.

Additionally, it seems prudent to fetch from the remote prior to checking the
status, to see how much divergence has taken place.