feat: extend fishing concept with jank
I don't know what I'm doing... lol

I made the "rod" show up and then I tried using a dampingspringjoint to have it
track the mouse, but that was too weird so I made it very subtle.

I also reworked the chain to make it more like a fishing line w/ a bobber.

I tried to make the pole stay in "front" of the player's movement, but it
doesn't move the other bits with it and it doesn't stay as I'd expect it to :/

Overall, fun to mess with still :)