feat: Trolley initial bottom panel listing controls
Unfortunately the bottom panel scene does most of the work for now, b/c
it's a better dev loop than editing code in a tool-autoload (which seems
to require a full project reload when updated to run in the editor).
Maybe this isn't a problem for working with the built-in editor - i
couldn't find a clear way to force the reload yet - if anyone knows of
one, plz lmk.