feat: actionHint component consumed by actionDetector
A tweak that prevents actionDetector actors from needing to impl this
display/hide logic every time - tho they still need to add and place an
action_hint on the actor.

There's an instinct to add this action_hint to the actionDetector, but
for now there are cases where the player wants to use it - specifically
for a sword attack when there is something to hit. Should that be
absorbed and impled via the action api?