feat: action_hint supports input_action label
The action_hint now attempts to find an input_action and keys assigned
to it from the project input map. if none is found, the passed text is
used directly, to support one-off keys (in particular, supporting groups
in action hints, like "WASD" for "move", which is several input maps.
This could be cleaned up at some point, and hopefully shared across
controller types.)