fix: avoid loading the current scene while saving
Elevators pointing to other elevators in the same scene were causing the
editor to crash - seems to be because of an underlying state problem
when reading from a packedscene's state while also trying to save that
scene. woof!

The good news is that we now know how to work with packed scenes more closely.