refactor: elevators store dest area name, not path
Also updates Util.packed_scene_info to set scene_file_path if the path
is used as input. This is necessary to load the next room when using
Mvania elevators, tho maybe there's a better way to work around that. We
could instead lookup the path in MvaniaGame using the area name, but
that'd be duplicating Hotel's purpose a bit - it's annoying we can't
reliably get the path from Hotel, which is already our in-game state db.