feat: introduce Hotel.register()
Cleans up the hotel api slightly - supports (requires) a
`check_out(data)` and `hotel_data() -> Dictionary` function for
restoring nodes from their hotel data and creating a new dict to be
stored from that data. Should be called from _ready().

In check_out(data), be sure to not overwrite local settings if no value
is passed from hotel! The intent here is to make it easy to restore the
state from a previous hotel_data call when re-creating a node, but the
first _ready() call needs to be careful. Perhaps it could skip the
check_out(data) call entirely...?