refactor: break MvaniaGame into HatBot, Game autoloads
The game and player life-cycle details are now handled by Game.gd, which
can hopefully be extended to run as the main game autoload for all dino games.

HatBot takes on more HatBot specific things, like the area and player
scenes, plus gluing some Game.current_game usage to the MvaniaGame. We
will make HatBot a DinoGame next, so the Game.gd use of it is
repeatable.

MvaniaGame will be renamed soon. For now it has a few helpers that seem
reusable enough: load_area, travel_to, get_spawn_coords, and
update_rooms. All of these could be implemented in HatBot - the question
is how much to implement in there vs leave in a reusable something or
other. We'll have similar questions for the MvaniaArea and MvaniaRoom
impls shortly.