feat: play zones in unmanaged mode
Restores individual zone scene testing! The zones have some setup
helpers that run in _ready() to ensure a current game in Game, a current
zone in Metro, and to spawn a player (if one isn't already spawning).

I introduced then fixed a bug here related to unmanaged travel - if you
travel to a new area, it should always take you to the other end of the
elevator. the side-effectful nature of the zone's spawn_path in
player_spawn_coords() was causing the spawn_path to be cleared before
the Game._respawn_player eventually looked up player_coords. This was
b/c I opted to pass spawn_coords through the maybe_spawn_player funcs.
Instead, we pass a function that will later be called. In the traveling
case, the unmanaged helper never fires (b/c we're already respawning due
to the navi-change-scene signal-listener), but the side effectful coords
calc was then clearing the spawn_path, so the player was spawning at
whatever dev-spawn-point was still in play.

Perhaps we should allow the zones to be responsible for spawning the
player, rather than doing it from the game when navi changes the scene?
I'd like the Game to be able to manage the player independent of the
game-impls/zone-impls, but maybe they're already coupled...

This also moves the player and zones to booking with Hotel in their
enter_tree() method, and removes the player booking from the game
register function. I'm not sure if this makes sense or not, so,
something to keep an eye on. I tried to remove the full game register(),
but if you don't register zones, the elevators don't work (b/c there's
no scene in hotel to fetch and load when traveling). Moving the zone
book to enter_tree() resolved a bunch of warnings from hotel for all the
children calling register/check-in/check-out.

I don't want metro or the zones to depend on the game impls (e.g.
hatbot), so instead Game.gd does - this could happen in a separate
autoload (Games.gd ?), but there's no circular-dep yet, so this might
last. Game.ensure_current_game() should let the zones be tested
independently without too much fuss - just need to impl
DinoGame.manages_scene for each game.