feat: include notifications in fallback hud
Refactors the Hood.ensure_hud to allow the fallback hud to include
components that depend on Hood. The lesson is that an autoload can't
preload a scene that depends on the same autoload (circular dep).

Now we get notifs for free with a simple Hood.ensure_hud() in any
node (likely the player).