feat: gunner hud break-the-targets data via Hotel
Adding the 'quests' as a root group to make this work for now, but we
should probably improve the key building to 'just-work'. Not sure how
just yet - can we determine the place in the tree at 'book' time?
probably not from the scene_file_path, but we probably can in
_enter_tree().

Also pretty brittle - using strings in hud to pull hotel keys/data that
other nodes checked in. could instead grab data or a whole component
from the quest node? Can we grab actual nodes from hotel?