feat: Hotel.book supporting node
Hotel.book now handles a node, and is expected mostly from
_enter_tree(), unless you are pre-emptively passing paths to a games
maps to support connecting things like elevators.

Adds dungeonCrawler player to hotel to test it.

This is big b/c otherwise Hotel wasn't working unless the nodes existed
in a pre-registered 'root' scene, which is not a limitation we care to continue.