feat: grab and throw sheep
Impled a quick Util.get_(opts, key, default) so that dictionaries with
nil values on keys don't get set. So don't use nil to overwrite values
if you're going to use this function!

Also add support for a maximum distance to actions. it might actually be
a minimum distance? but you have to be 'below' it for it to be true, and
i could see some minimum distance coming into play later on. The default
proximity check is just that the actor is in the actionArea itself, so
this lets us be more specific per action.