feat: bullet/spawnPatterns on wolf, home by group
Moves the bullet/spawnPatterns back to the wolf scene, drops the
HerdBulletPatterns autoload. The bullet/spawnPattern ids are updated to
per-instance ids in _enter_tree(), which lets us manage the wolves'
patterns on the wolf itself, which is nice. Hopefully this won't lead to
performance issues, but that should be capped somewhat by the pool size.

Moves the homing logic to use groups instead of using
Spawning.change_property to update a node_path. This is a bit cleaner
and probably feels better from a gameplay perspective anyway, and
removes some of the need for a `target` reference on the wolf itself.