fix: check if homing_target has been freed
Another BulletUpHell crash - after freeing the player, this line was
crashing. The crash is strange - it shouldn't happen b/c of this `get`,
it should happen later when we try to use the value. For whatever reason
using the dictionary[key] syntax was fine.

Feels like a bug in gdscript? Either way, we probably want to check if
the instance is valid before continuing in this conditional.