feat: add sounds, screenshake to misc states
Could debate where to put these for a long time! Until the animations
get added, we don't know the complexity of the timing for sounds...

There's also the question of how to handle different sounds for
different players, and then of the names of the sounds themselves - they
are quickly drifting. I'd hope it would be easy to interact with the
sounds between the DJ TurnTable and gdfxr. Ideally the existing sounds
could easily be edited, cloned, renamed, and still included in DJZ.