refactor: move BEU machine and player into beehive/beatemup
Moving toward including ready-made controllers and AI in beehive.

I intend to work with this beatemup controller in multiple games going
forward, so opting to move this across the game/addon boundary early to
start feeling it out. This makes room for SuperElevatorLevel-specific
things to happen with a proper warning/resistance to muddying the
game-agnostic addon code.