fix: lots of is_instance_valid in BEU states
This is ugly as hell, and still doesn't really seem to catch everything.
It looks like dictionaries that refer to freed instances can't be
accessed with 'get' as well, which makes things terrible. probably
that's a bug or something i'm doing wrong... this Util.get_ helper
didn't quite catch it either.

Any who, the player seems to be respawnable without crashes now.