fix: don't pause all metro rooms when the player isn't in ANY
starting outside a metroRoom was causing _no_ rooms to ever be found,
because they were all paused, which led to none of their room_entered
signals firing. This fixes that and another bug, where the player global
position was used when it might not have fully entered the room yet. now
we support a get_rect on the player and a rect.intersects() chec.

Also adds a directional light and some light-occluding tiles