feat: add actions api to player
Extends the actions API to support actions passed in from the player
itself. The ActionDetector script ought to be broken out into an Actions
autoload that can register an actor and remove some of this
connecting/boilerplate.
could even return signals to .emit() on for execing and recalcing
current_action.