fix: has_player calc using same room.contains_player()
DRYs up a room-contains-player calc. We could maybe use the roombox for
this instead... i think the logic ends up being the same. Just note that
rooms may be paused when we want to run this, which could affect the
roombox signals, tho not necessarily the roombox's geometry/collision
calcs.