feat: focus seems to be working
Any scene used with Navi.nav_to can impl `set_focus` to handle setting a
particular focus after navigation.

A fallback implementation attempts to find a focus if set_focus is not a
method on the current scene.

This _might_ compete with grab_focus() in a scene, which could be called
by a control node - maybe we want to defer this and check for a focused
node via viewport.gui_get_focus_owner first.

Either way, it's nice to be able to play through a few games with just
the controller now!