feat: add 'Dino Menu' option to relevant menus
This won't show up in exported games, unless it's the 'dino' build, but
it should always show up in the editor-based builds. Feature tags make
this fairly simple!

Also extends Navi.find_focus to accept a passed scene, which helps
support the fallback focus-first-button approach for
non-set_focus-implementing pause/win/gameover overwrite scenes.

Probably still a menu or two left to fix this on - maybe harvey's
timeout screen... i'd really like a shared fix for that, not handling it
in every case individually.