fix: support menu.show() focus cases
Ideally this would get solved in a built-in way, not a
don't-forget-to-call-Navi.find_focus(menu) after menu.show() way.

Maybe menus always get navi-registered and shown through a navi helper?

Otherwise, how to detect that there's nothing in focus after a
menu.show() happens?

I'd really like hitting tab/up/down/some ui button to notice that
nothing is in focus when trying to find a neighbor, and then give an
opportunity to focus on the current scene - tho in this case, we'd also
need to know that there's a menu overlay shown on top of it. Is there a
way to know what scene/menu is currently on-top? Maybe navi needs to
manage all the menus and the current scene. that might resolve the
pause-while-on-dino-menu oddity.