feat: move DinoGame, Game to core, invert deps on game autoloads
Also moves some deps/logic from Debug.gd to Trolley.gd, to make Debug.gd
easier to autoload (it's the first autoload, so should not have deps
like Cam/Trolley/Hood). Trolley would ideally also not depend on
Cam/Hood.

Moving DinoGame to an addon to make it accessible to BeatEmUpCity, which
a first consumer of the dino addon suite.