fix: only use supported colors in pretty printer
Fixes a long-annoying bug in a pretty nice way. Reading the print_rich
doc-string, it turns out supported colors are converted to ansi codes
when output to stdout, which means the test and stdout in the
terminal/emacs buffer are now MUCH more readable (not full of noisy
color labels and reasonably colored). The color choices are limited, but
we could make it more readable with some bgcolor/fgcolor usage.