feat: adding room above last room
Basic version of adding a room above the last one. Probably not quite
the right api yet - this supports passing a side to add the next room
too, but we might want to search the room options for one that fit the
last room's open sides, and start getting into rotating the rooms.