feat: now adding player + enemies
Not too bad! Didn't even need to subclass BrickRoom here, which I'm
thinking is a bad pattern. Not sure if BrickRoom should even have an
instance... vs just creating and passing along node2ds or some
non-specific thing. It's really about the tiles and entities - the tiles
are being promoted to the level, and the entities probably should be as
well.