fix: don't set tree root as current_scene
Finally learned how to read this error - I was setting the root as the
current_scene, which is a no-no. For now returns to the previous
algorithm, setting the root's last child as the current_scene, which
should generally be right? Hopefully this removes another error failing
the start-every-game test.