feat: woods levels generating via BrickLevelGen
Not 100% about subclassing BrickLevelGen every time, but maybe it makes
alot of sense?

Starting to implement level logic, in this case just finding the last
room and connecting to it's signal.

Introduces roomboxes on Rooms, which requires even more assignment of
node children owners to keep (for the required collisionShape2Ds).

Extends BrickLevelGen to accept nodes for entities/rooms/tilemaps, to
DRY up the node handoff from gen to level. Could even create and add
these on the fly to the gen's parent or owner.