feat: add some harvey levels, adjust box layouts
For these grid-based levels, rather than push every entity down to the
bottom of it's coord, we expect to align 0,0 on an entity with the
grid's coord (top-left). Further adjustments can happy per entity in a
passed setup(ent) helper. The baked-in helper made sense in a platformer
context, but not a top-down one.

Another adjustment is making seedboxes accessible from all sides, which
makes more top-down sense, and avoids the rotation complexity that the
delivery boxes took on.