feat: spike levelgen impl working, including top/bottom portal
Kind of a hack, but these are the escape hatches i want to support.

Could refactor towards a simpler api, something that automagically
reparents/rearranges cross-dep entities like this. But, realizing that
this feature could be impled via the roombox makes me wonder if it's
better to keep things simple - i.e. find some other way to express it.
we'll see how it goes as we add more use-cases - a simple "setup" called
on the final generated output might be a better/simpler place to do some
of this work, tho right now we're benefiting from letting the rooms do
the positioning work.