feat: first game_mode, drop a bunch of game singletons
Introduces a new concept, dino game modes. This is a sub-category of
DinoGameEntities right now.

Drops a bunch of unused game_singletons, and refactors Game.gd to create
a DinoGame on the fly when a singleton isn't specified. DinoGames still
use 'singleton'-like behavior via the Game.gd autoload.

Perhaps the Game.gd autoload should also go away or otherwise lose it's
state? Right now it'd be problemmatic to run two games at once. So far
it seems like the encounter/level refactor we're in will change the way
this works.