addon: adds custom-scene-launcher and ports it to godot 4
Imports: https://gitlab.com/godot-addons/custom-scene-launcher and
manually updates most of it for godot 4. It sort of works - not sure
what the pin button is supposed to be doing. Will probably strip it down
at some point - not sure why it's reading/writing from a config at all,
for example.

Either way, a bit of a boon for now, as we can manually select a scene
and play it regardless of which scene is focused.