fix: pause menus on layer 10, process_mode 'always'
Figured out a long-time issue where clickign on the pause menu wasn't
doing anything. Finally diagnosed that it was the player HUD blocking
it... but that's not right, is it? It's the canvas_layer! We don't need
to re-order the scene_tree, we need to bump up the layer of the pause
menus so they get that always-in-front treatment.