feat: initial test coverage for ssplayer add_weapon(ent_id)
Getting some coverage in place over DinoPlayer weapon crud before drying
it up. Right now there's duplicate logic in SSBody and TDBody, but the
code is (nearly) the same.

Only tweak here was a null-check in a weapon's activate(), which crashes
in the test because the actor is set at 'ready()' time. Maybe it
should/could be set before then? enter_tree()? maybe that makes sense?