wip: potential direction for some weapons helpers
But it feels like we should bury the weapons state in a node or
object... it'd cut down on the exposed api (no need to pass the
'weapons' state around...), but it's quite anti-functional.

A different direction might be passing the player into a
weapons system call instead of asking the player for weapons data.