wip: first pass at a spike + target level
Doesn't run yet because it doesn't know what kind of player to spawn -
need to refactor away from depending on dino-game for player type.
Perhaps we can move to letting dino level handle it.

DinoGame as a requirement is a bit overloaded - we'll want to gather
whatever those deps are on the fly, and eventually we'll want to change
player type on a per-room basis, not per level.