feat: drop Player (P) autoload, use Dino + PlayerSet
Relying on Navi in the PlayerSet for a node with access to the tree, so
we can find the spawn_point when spawning.

I abandoned some extra logic and 'spawning' flags that were in place to
prevent extra players from being spawned - i don't think we need them,
and they should really be more of a smell than anything.