feat: very rough but working classic mode
Implements a classic mode based roughly on Roulette mode, but generating
levels via LevelDefs and DinoLevel's create_level instead of
DinoGameEntities. We'll need to expand the 'stages' from simple
level_defs to level_defs + opts to mix in (varing tilesets and level
layouts). Tho really i expect classic to be ~10 unique levelDefs, so
maybe that won't be necessary.

I'm pretty happy with this - we'll be able to dry up all the game
components and recreate the 'games' as lower-key levelDefs, which is
roughly a levelscript set of rooms + legend paired with a
gen_room_opts() function.