refactor: remove deferreds from BrickLevelGen
I don't want to wait for a tick or 'ready' or whatever nonsense for a
level to be generated unless it's truly necessary - this refactors such
that DinoLevel.create_new_level() returns a completely generated
level_node, and a player can be spawned on it by passing the node to
Dino.spawn_player(level_node).

This is setting up a showdown between the generated rooms and the metSys
opinion... not sure what will survive.