refactor: composable RoomInputs, basic tests
Introduces RoomInputs, a class full of static functions for
combining/merging tile_sets, entities, and room shapes.

RoomInputs do the heavy lifting when defining VaniaRoomDefs to be
generated.

The VaniaRoomDef acts as a 'RoomData' object for each VaniaRoom during
generation and while the game is running (for getting neighboring room
data, for example).

RoomInputs are intended as the data object to be composed together and
targetted by level-design apis. some of it's static helpers present one
field (like a single tilemap), others present them in
combination (entities + room_size). These will grow as we add props,
backgrounds, npcs, etc.

This also starts to write some basic test coverage for
vaniaRoomDefs/inputs and VaniaGeneration, which has been desperate for
it. The basic cases are not much, but should make it easier to reach for
when debugging/extending - e.g. supporting concave room shapes is a
headache, tests should make it much easier.