wip: test covering gen-time door creation
the doors (and walls) are recalced at _ready time right now, but ideally
we'd update them at generation time - this requires some
neighbor-awareness, and updating/repacking neighbor rooms when a room is
added/removed. I'm not sure how costly this is... but we probably should
only update affected rooms, not all of them.