refactor: Quests via manager node instead of autoload
The Quest manager now checks it's parents children for quests, and
subscribes to add_node events on the tree to trigger updates.

This could present an issue with 'totals' moving around, and the
approach has some clear performance/ugliness, but the general flow seems
like the preferred one. Next, to integrate into vania rooms a bit.