feat: action areas self-register
Similar to actionDetectors, actionAreas now register themselves when
their parent is 'ready'. It looks for an ActionHint on the parent and
expects some 'actions' to be defined.

We should probably warn if no actions are defined but the area/detector
exists and aren't registered yet.

`register_actions` can still be used to set actions on custom
areas (e.g.. the coin door has separate front/back action areas)