fix: restore thread-based level gen
Pre-calcs some neighbor data that was pulled from Metsys - the metsys
data uses signals to updates its state, which unfortunately complicates
doing that work on a thread (b/c signals/deferreds are awaited on the
main thread). There might be some clever locking/semaphore workarounds,
but instead we precalc the neighbor data that we were building from
metsys - ideally we'd be able to run without it.

Adds some awaits to tests that use MetSys's state to sanity check the
data.