test: coverage for vaniaGame threaded level gen
Initial tests for vaniaGame's threaded level gen, including removing the
VaniaGame node from the tree while the thread is running.

Lots of dodging create_tween when not in the tree - maybe ought to wrap
that/try for a null-safe/nil-punned tween api. sort of happening in the
Anim class static funcs for now.