fix: don't apply gravity to wall crawlers
Wall-crawlers no longer have gravity applied (in idle and run states).
this fixes crawlers sticking to the ceiling and wall starting positions,
tho they need to be overlapping/_very_ close to the walls for
is_on_wall/ceiling/etc to detect the wall - could be better to use a
raycast to find the nearest wall and move toward it.