feat: activate/deactivate room lighting/bg features
Now that we're rendering multiple levels at a time, these room features
are conflicting w/ each other - this prevents the overlap by
deactivating other rooms, and activating only the current room's
lighting feats.

This also drops an unused HUD component and serializes the tower, woods,
and mountain mapdefs as individual files.

There's some oddity in the quest completion still - i was able to
complete the tower with an enemy and target remaining - for targets, it
might makes sense b/c some target configs present multiple nodes for a
single entity - for the enemies, i'm less sure...

Also some trouble with enemies getting stuck in walls - a 'skipped'
ceiling is clear, but then when you go through it (as a vertical door)
some gets filled in (the floor of the room above it). For now i may have
to disable the skipped ceilings on the tower level.