feat: support awaiting scene exit before adding new one
The phantom camera was compaining about too many hosts, b/c we were
adding the new root scene before the previous one was removed. This
should prevent other issues, like spawn_player finding and using the
previous scene's spawn_point.

We might want this as the default, and to opt-in to skipping this await instead.